[#3] TF : updating Stopwatch and almost finished basic audio recorder/player

+++ app/src/main/java/com/abstractx1/mydiary/record/Recorder.java
@@ -28,7 +28,7 @@ public class Recorder {
         stopwatch.stop();
     }
 
-    public int getRecordingDuration() {
+    public float getRecordingDuration() {
         return stopwatch.getElapsedTime();
     }
